+++ src/Backend/Items/spMacroItem.ts
@@ -1,7 +1,0 @@
-import { CompletionItemKind } from "vscode";
-
-import { FunctionItem } from "./spFunctionItem";
-
-export class MacroItem extends FunctionItem {
-  kind = CompletionItemKind.Interface;
-}