+++ src/Keyvalues/registerKVLinter.ts
@@ -1,33 +1,40 @@
-import {
-  TextDocument,
-  workspace as Workspace,
-  Range,
-  Diagnostic,
-  DiagnosticSeverity,
-} from "vscode";
+﻿import * as vscode from "vscode";
 
-import { parse, SyntaxError } from "../Parser/kvParser/kvParser";
-import {
-  KeyValue,
-  ParserOutput,
-  ParserRange,
-} from "../Parser/kvParser/kvParserInterfaces";
-import { kvDiagnostics } from "./Linter/compilerDiagnostics";
+import { parse, SyntaxError } from "./kvParser";
+import { KeyValue, ParserOutput, ParserRange } from "./kvParserInterfaces";
+
+// Register and export the DiagnosticsCollection objects to be used by other modules.
+export const kvDiagnostics = vscode.languages.createDiagnosticCollection("kv");
+
+export function registerKVLinter(context: vscode.ExtensionContext) {
+  context.subscriptions.push(vscode.languages.createDiagnosticCollection("kv"));
+  context.subscriptions.push(
+    vscode.window.onDidChangeActiveTextEditor((editor) => {
+      if (editor) {
+        refreshKVDiagnostics(editor.document);
+      }
+    })
+  );
+  context.subscriptions.push(
+    vscode.workspace.onDidCloseTextDocument((document) => {
+      kvDiagnostics.delete(document.uri);
+    })
+  );
+}
 
 /**
  * Lint a Valve Key Value TextDocument object and add its diagnostics to the collection.
  * @param  {TextDocument} document    The document to lint.
  * @returns void
  */
-export async function refreshKVDiagnostics(document: TextDocument) {
+export async function refreshKVDiagnostics(document: vscode.TextDocument) {
   await null;
 
   // Check if the setting to activate the linter is set to true.
-  const workspaceFolder = Workspace.getWorkspaceFolder(document.uri);
-  const enableLinter = Workspace.getConfiguration(
-    "sourcepawn",
-    workspaceFolder
-  ).get<boolean>("enableLinter");
+  const workspaceFolder = vscode.workspace.getWorkspaceFolder(document.uri);
+  const enableLinter = vscode.workspace
+    .getConfiguration("sourcepawn", workspaceFolder)
+    .get<boolean>("enableLinter");
 
   // Stop early if linter is disabled.
   if (!enableLinter || document.languageId !== "valve-kv") {
@@ -41,7 +48,7 @@ export async function refreshKVDiagnostics(document: TextDocument) {
     parsed = parse(document.getText());
   } catch (e) {
     if (e instanceof SyntaxError) {
-      const range = new Range(
+      const range = new vscode.Range(
         e.location.start.line - 1,
         e.location.start.column - 1,
         e.location.end.line - 1,
@@ -49,7 +56,7 @@ export async function refreshKVDiagnostics(document: TextDocument) {
       );
 
       const msg = e.name + " " + e.message;
-      const diag = new Diagnostic(range, msg);
+      const diag = new vscode.Diagnostic(range, msg);
       kvDiagnostics.set(document.uri, [diag]);
     }
     return;
@@ -57,9 +64,9 @@ export async function refreshKVDiagnostics(document: TextDocument) {
   kvDiagnostics.set(document.uri, lookForDuplicates(parsed.keyvalues));
 }
 
-function lookForDuplicates(keyvalues: KeyValue[]): Diagnostic[] {
-  const map = new Map<string, Range[]>();
-  let diagnostics: Diagnostic[] = [];
+function lookForDuplicates(keyvalues: KeyValue[]): vscode.Diagnostic[] {
+  const map = new Map<string, vscode.Range[]>();
+  let diagnostics: vscode.Diagnostic[] = [];
   for (let keyvalue of keyvalues) {
     const range = parserRangeToRange(keyvalue.key.loc);
     const key = keyvalue.key.txt;
@@ -76,23 +83,27 @@ function lookForDuplicates(keyvalues: KeyValue[]): Diagnostic[] {
     if (prevRanges.length === 1) {
       // Add the first diagnostic.
       diagnostics.push(
-        new Diagnostic(
+        new vscode.Diagnostic(
           prevRanges[0],
           "Duplicate object key",
-          DiagnosticSeverity.Warning
+          vscode.DiagnosticSeverity.Warning
         )
       );
     }
     prevRanges.push(range);
     diagnostics.push(
-      new Diagnostic(range, "Duplicate object key", DiagnosticSeverity.Warning)
+      new vscode.Diagnostic(
+        range,
+        "Duplicate object key",
+        vscode.DiagnosticSeverity.Warning
+      )
     );
   }
   return diagnostics;
 }
 
-function parserRangeToRange(range: ParserRange): Range {
-  return new Range(
+function parserRangeToRange(range: ParserRange): vscode.Range {
+  return new vscode.Range(
     range.start.line - 1,
     range.start.column - 1,
     range.end.line - 1,