+++ src/spUtils.ts
@@ -1,5 +1,4 @@
 ﻿import {
-  CompletionItemKind,
   Location,
   MarkdownString,
   Range,
@@ -8,11 +7,6 @@
 import { URI } from "vscode-uri";
 import { existsSync, lstatSync } from "fs";
 import { resolve, extname } from "path";
-
-import { SPItem } from "./Backend/Items/spItems";
-import { MethodItem } from "./Backend/Items/spMethodItem";
-import { PropertyItem } from "./Backend/Items/spPropertyItem";
-import { VariableItem } from "./Backend/Items/spVariableItem";
 
 /**
  * Parse a Sourcemod JSDoc documentation string and convert it to a MarkdownString.
@@ -62,8 +56,9 @@ export function findMainPath(uri?: URI): string | undefined {
   const workspaceFolder =
     uri === undefined ? undefined : Workspace.getWorkspaceFolder(uri);
   let mainPath: string =
-    Workspace.getConfiguration("sourcepawn", workspaceFolder).get("MainPath") ||
-    "";
+    Workspace.getConfiguration("SourcePawnLanguageServer", workspaceFolder).get(
+      "mainPath"
+    ) || "";
   if (mainPath === "") {
     return undefined;
   }
@@ -103,30 +98,42 @@ export function locationFromRange(filePath: string, range: Range): Location {
 }
 
 /**
- * Check if an item should be a constructor instead of a methodmap, given the line it was found in.
- * If it is a constructor, return the correct item. Otherwise, return the original object.
- * @param  {SPItem} item  The item to check.
- * @param  {Range} range  The range of the match.
- * @param  {Map<string, MethodItem|PropertyItem>} methodsAndProperties  The methodsAndProperties of the parser.
- * @param  {string} line  The line at which the item was matched at.
- * @returns SPItem
+ * If needed, migrate the settings of the user to use the LanguageServer.
  */
-export function checkIfConstructor(
-  item: SPItem,
-  range: Range,
-  methodsAndProperties: Map<string, MethodItem | PropertyItem | VariableItem>,
-  line: string
-): SPItem {
-  if (item.kind !== CompletionItemKind.Class || line === undefined) {
-    return item;
+export function migrateSettings() {
+  const smHome: string =
+    Workspace.getConfiguration("sourcepawn").get("SourcemodHome");
+  const optionalIncludeDirsPaths: string[] = Workspace.getConfiguration(
+    "sourcepawn"
+  ).get("optionalIncludeDirsPaths");
+
+  const includesDirectories: string[] = Workspace.getConfiguration(
+    "SourcePawnLanguageServer"
+  ).get("includesDirectories");
+
+  const oldSpcompPath: string =
+    Workspace.getConfiguration("sourcepawn").get("SpcompPath");
+
+  const newSpcompPath: string = Workspace.getConfiguration(
+    "SourcePawnLanguageServer"
+  ).get("spcompPath");
+
+  if (
+    (includesDirectories.length == 0 && smHome) ||
+    (!newSpcompPath && oldSpcompPath)
+  ) {
+    Workspace.getConfiguration("SourcePawnLanguageServer").update(
+      "includesDirectories",
+      Array.from(new Set([smHome].concat(optionalIncludeDirsPaths))),
+      true
+    );
+
+    if (oldSpcompPath && !newSpcompPath) {
+      Workspace.getConfiguration("SourcePawnLanguageServer").update(
+        "spcompPath",
+        oldSpcompPath,
+        true
+      );
+    }
   }
-  const subLine = line.substring(range.start.character);
-  const re = new RegExp(`^${item.name}\\s*\\(`);
-  if (!re.test(subLine)) {
-    return item;
-  }
-  return (
-    (methodsAndProperties.get(`${item.name}-${item.name}`) as MethodItem) ||
-    item
-  );
 }